refactored into separate reusable behaviour class
